R# Rename: Memory -> InMemory
